update mocha grapher tests for new dataPoint array
The structure of the dataPoint arrays passed to the grapher
have changed. The Mocha tests for the graph controller are
now updated correctly to reflect this.

New top-level dataPoint array structure:

  [ <channel0>, <channel1>, … <channeln> ]

Each channel consists of an array with a sequence of 0..n
two-element arrays of x,y data pairs.

This is the form used when accessing the grapher with
the the new addPoints and resetPoints methods.

This is also happens to also be the form used internally
by the grapher.

A different form or array structure is used for the
methods addSamples and resetSamples. These methods are no
longer used in the graph controller.