grapher: a selected point is draggable along plot
If options.dataChange is false then a drag left or right
on a selected point will shift the actual point selected
left or right.

A second mousedown click on a selected point will un-select it.

TO: allow multiple points to be selected. A user with a
keyboard could use shift-click … not sure what he equivalent
is on a tablet/touch interface.