Create PluginController, and the ability to add md2d plugins.
A plugin is a set of related functions and changes to the md2d arrays, which
would normally be in md2d.js, but can be extracted out and optionally enabled
or disabled.

For example, all functions related to Quantum Dynamics can be extracted out, so
that they can be found an a single meaningful place, and so that they do not
need to be invoked when a model is not using quantum dynamics (and so md2d doesn't
need a large number of "if (quantumDynamics) { ... }" statements).

Since this is code that would otherwise be directly in md2d.js, the plugins are
"privileged" and can update atom tables and do other things that can be freely
done within md2d.js.