Add initial stub of QuantumDynamics plugin
QuantumDynamics plugin initially just adds a new array to the atoms array,
"excitation", which is an int representing the current level of excitation
of an atom, where levels are defined in a new not-yet-existant property,
"elementEnergyLevels", to be defined in the model.

Per Classic MW, we currently (and for the known-future) make the simplification
that each atom only has one single electron that can be excited to a new energy
level, and so we can safely model this as a single int per atom.