Add quantum dynamics props to metadata
As part of having quantum dynamics being its own self-contained
plugin, it seems desirable not to need to update all other model
json files (or their serializations) with default QD property values,
such as energy levels or excitation.

For this reason, the QD properties do not have a default value (or
they would be serialized).

Another option was to define the new properties in the plugin itself,
and add them to to metadata dynamically as needed. However, so further
restructuring of the metadata would be required to allow that.

A consequence of this is that when we initialize the model, we have to
check if a property exists in the atoms table, and not just assume that
if it's in the metadata then it must be in the table.

We can also consider other ways to handle this issue.