QD: Add excitation after collisions; call within intgration loop
Quantum dynamics plugin will now excite an atom if two atoms are
close enough to each other and their relative KE is above that required
to reach the next electron energy state.

If an atom gets excited, the velocities of the two atoms will drop,
resulting in a drop in KE. This will also result in a drop in total
energy, because the potential energy of the electron state isn't
added to the model's potential energy.

The excitation algorithm is functionally identical to the one used
by Classic MW.

MD2D calls "performActionWithinIntegrationLoop" as the last action
within the loop, and passes the current neighborList (for quick finding
of close neighbors within QD). Any other plugin that needs to act within
the integration loop can also respond to this action, and will get called
automatically. If they need new arguments passed in, they can easily be
added to the object argument without needing to update existing plugins.