De-excite atoms during collision; possibly increase KE
When atoms collide, if one atom has an electron in an excited
state it will dexcite. It will then randomly either emit that
energy as an increase in the velocities of the two atoms (>KE),
or as a photon (not yet implemented), with a probability based
on the model settings.