Always excite an atom if one can be excited
Previously if both atoms had electrons that could be excited,
we randomly selected one and then tried to excite it.

Now we check the result of that excitation and try to excite
the other if the first one failed.

This prevents the situation where we have enough energy to only
excite one of the atoms, but we randomly picked the other, resulting
in neither atom getting excited.

[#48674803]