Prevent multiple QD operations during same collision
We maintain a list of the pairs of atoms that have just
had a QD operation, and only clear them from that list
once their collision has ended. This allows us to prevent
multiple QD operations occurring during the same collision.

Classic MW handles this slightly differently, by defining
a minimum-interval time during which the quantum engine
will not respond to events. This seems to be a slightly
cleaner way of achieving the same ends.

[#48674945]