Initial implementation of photon table and emission
QD plugin creates an empty table of photons, much like the
atoms table, which gets extended and filled in as needed as
photons are created. If they are created, this will be 
serialized to the model json.

Photons can currently only be created on collision. QD uses
Classic MW's default values for planck constant and speed
of light.

Engine's call to the plugin now also passes in dt, as we need
this to move the photons.

[#48675225]