Remove the 'quantumDynamics' model property.
Model properties should be primitive values.

Making the quantumDynamics object a property of the model isn't quite right because it's not a primitive value, so the property observing code doesn't detect changes. Also making quantumDynamics a property of the model caused it to be serialized via a "back door", rather than being explicitly serialized by passing the quantumDynamics metadata to the serialization module.

Note that this commit temporarily breaks a serialization test.