Shapshot phantom: Save html, use base_href, &  CSS
Instead of rewriting the html and css, we simply change the doc base.
We are saving both HTML and PNG files now, so that we can debug the html

Shapshot application.js: extract snapshot.js send css, send base_url

I removed the snapshot specific code out of application.js and into its
own javascript file.

Also: We now send all the stylesheets in the page, and a base url.