Implement LabModelerMixin - tool for new model types
[#51900021]

This module should be useful for all Lab model types. It combines
PropertySupport, ParameterSupport and OutputSupport, as well as
defines builtin properties using provided metadata and units
definiton.

Of course this can be done manually, but this provides easy, ready-to-
go configuration, which is used by all model types at the moment.

When a model doesn't need all these features (mixins) or it needs
something special - this still can be a hint how to do it.