get all zone records when listing AWS servers
Display hostnames when running: thor cloud:list even
when the zone record list is large.

Just requesting all results in a paginated type of
response from Route53. Using the all! method in Fog
actually gets all the records. 

This is inefficient. It should instead be possible
to get a DNS record from Route53 with fog by specifying
a record type of "A" and the right IP address.