interactive migration: add type to model references
This is just a very first early hack to run a migration
changing the schema for Interactives.

Model references in Interactives now include a type attribute.

There are a number of useful reasons for apps that manage
Interactives to know about model types used in an Interactive
without having to load the model.

The next commit will include all the changed Interactives
and updates in the metadata.