Interactives: optional importedFrom attr
Interactives now have an optional importedFrom attribute.
If present the value indicates a path to the HTML or CML
page the Interactive was imported from.

Model references in Interactives also have an optional
importedFrom attribute. The value indicates a path to
the MML or E2D file the model was imported from.