Guardfile: create 2 group: 'build' and 'test'
Starting guard normally will continue to run all watch tasks
in response to file-change events.

 $ bin/guard

To run just the build steps in response to file-change
events start Guard with the group 'build':

 $ bin/guard --group build

The Mocha and Vows tests will *not* be automatically run
when changes are made to files in src/lab and selected
files in src/helpers.

However if you make a change in a test file itself tests
will be run.