eliminate "ghost" photons on tick history restore
[#48675251]

Without this change, restoring from a state which had a smaller photons table fails to clear the vx and vy entries of photons that existed in the prior state, but which are in the newly extended region of the photons table.

The alternative to padding the restored photons arrays is to stop iterations over the photon table when numPhotons non-empty slots are observed, but the approach of clearing the arrays seems much less error prone.