mocha, phantomjs updates
make public/test

Creates and copies resources to the ./public/test dir
and uses mocha-phantomjs to run the tests in test/test1.js
in the generated html file: embeddable-test-mocha.html

FIXME: not sure why the tests are run before the
async beforeEach completes when run in the browser.

This cause a failure in the second test.

After running: make public/test

Load this page in a browser:

http://localhost:9292/embeddable-test-mocha.html#interactives/samples/1-oil-and-water-shake.json