Add run # to sensor data collector
Note that because we will be removing the "variables" section from the interface, I also removed the handling of "metadata" (variables) from the export logic. In turn this made it possible to replace the original/custom dgExporter with the Lab version, as bundled in the Lab.importExport.js build product.

[#49151437]