add another hack to interactive serialization
FIXME:
This fixes the failing interactive serialization mocha
tests … but we need something more robust.

helpTips have a defaultValue of an empty array … but
they are not required as part of the JSON form for an
interactive.

But when we check to make sure that the serialization of
an instantiated interactive matches the original JSON
without this hack helpTips is added … and the serialized
JSON doesn't match the original.