Fix semantic layout stop algorithm condition
[#54749558]

The culprit was that we were expecting that layout fits interactive
container +/- 1px, but also that content width is always smaller than
container width. It was causing that this condition was never met,
even when content was 0.001px wider than container. Fix that
and tolerate 2px of difference (invisible for users anyway).