LabModelerMixin: "willReset" event & method
The willReset method mixed in to models triggers the willReset event, which notifies observers that it the model is about to reset. This allows observers to inspect the model's state before that state disappears. (For example, the export controller wants to know about the state a model was in when a student or other user resets it).

IS [#54801606]