Don't automatically sort params to left of export.
The intent was to put properties that are per-run "inputs" to the model to the left of properties that are per-run "outputs" of the model. But there are legitimate cases for putting builtin properties in the exports, and this mechanism can't decide where to put them. Moreover authors should be able to rely on the order being what they requested, and are perfectly capable of choosing a sensible order.