scriptingAPI loadModel: pass onLoadScript not function
The optional second argument to loadModel must now be
the string form of the onLoadScript -- not the actual
function object.

Example:

"action": "loadModel('npPulling', 'start();');"

However passing an onLoadScript this way won't cause start()
to be run when the reload() function is run.

Ao also fix the reset button action by adding start() actions
to onLoad handers in the models specification.

Now pulldown components use the loadModel attribute in
their definition instead of using "loadModel('<model-id>')"
in an action.