Fix cell list reinitialization
[#55660284]

Previously we assumed that the model width and height are immutable.
However after recent refactorings, it's no longer truth. Model
dimensions are set a bit later in the initialization process and at
the beginning model is assumed to be 10 x 10. Cells ware using such
dimensions and they were never updated. Cells dimensions were
calculated incorrectly due to that.