important: don't expose 'model' object to scripts!
We don't want to discover months or years down the road that scripts in the wild, perhaps not even in our source control, depend on implementation details of 'model' and break when we try to refactor it. This is the main reason we have script sandboxing, in fact.

It's OK to make 'model' accessible as a global for console debugging; our script sandbox explicitly hides stray globals from scripts.