interactives controller getModel() accessor
This will allow us to eliminate the pattern in which the interactives controller passes a model object to every controller it creates when an interactive loads. This pattern makes it difficult to separate the code that renders components (which only needs to happen once) from the code that handles rebinding the components to the new model when it is loaded or reloaded.

Yak-shaving step toward IS [#56448478]