Move "performActionWithinIntegrationLoop" call before forces update
It seems to be safer, as otherwise e.g. chemical reactions plugin
was causing engine to diverge. I suspect that reason was that new
radial bond was added, however forces and velocities weren't updated
immediately, so bond was unnaturally squeezed or stretched. Now even
if a new bond is added, forces will be updated immediately.

Note that it also affects users actions. In fact all user actions
(or script actions) are executed between .integrate() calls. Not sure
what is correct place to modify model during the integration.

Perhaps we should always updated forces after changing model - some
operations do it, but not all for sure.