grapher: fix xDomain, yDomain methods
IS [#56344486]

Setting xDomain previously had no effect because the domain would be set from the argument, then updateXScale would reset the domain to xmin..xmax. Similarly for yDomain. Additionally the getter form returned a reference instead of a safe copy of the xDomain.

It appears that xmin/max, ymin/max are intended to be the sources of truth for the graph domain, and are updated when the user drags the graph. So use them always.