webgl checkbox: pass proxy interactiveController
[#57020306]

The Energy2D model adds a model dialog displayed by
clicking in the lighting bolt icon in the upper
left of the model.

Inside the dialog is a checkbox for enabling or disabling WebGL.

With the recent re-factoring of the way Interactive components are initialized the creation of this CheckboxController inside WebGL status dialog fails.

TODO: still need to find a way to call the modelLoaded
callback in this checkBox controller when the Interactive
is reset.