reorder elements below and above atoms according to layer position
This simplifies (and complicates) the layer structure in the md2d renderer. Instead of
having separate layer container for images, text, shapes, line etc with matching top
and below containers, there is now a container below and container above. Rendered
non-physical elements are appended into these containers and then sorted by
layerPosition.

MML test files with lines and textboxes now convert layerPosition, tests are updated.
A new test model and interactive is in place 'many-layered-objects.json' that
demonstrates how images, lines, and shapes can be interspersed on the above and
below layers.

This fixes the use of lines under atoms and above images in the Force Direction
interactives.