Defer initilizing sensors after sensor applet callback
This fixes a bug in Firefox on OS X where the reference to the applet is
made too early, so it generates a "x is not a function" error.

This error can still happen, though, if the JVM spins up in order to run
the applet (eg the first time you access the page).

PT [#56872730]