expose 'isUnexportedDataPresent' to scripts
Note this is exposed as a method rather than an observable property. Punting for now on what to do about potentially-useful properties like this that aren't specific to a particular model type and aren't explicitly specified by an interactive. (For example, we have no concept of "reserved property names" that shouldn't be used by an interactive. Also, perhaps this is an example of a property that should be defined on the interactive rather than on the model; right now, there is a single property namespace and it's owned by the model, but we already know that that won't be sufficient when we allow for multiple simultaneous models in an interactive.)