prevent error if user double-clicks the start button
The start button would disable properly and work fine if you single click the button.
There is code to monitor the state of the model to enable/disable the start button.
However, using an onPropertyChange observer was not fast enough to catch a
double-click, which I have seen many users do when pressing web buttons.