QD: Use Lifetime to decide whether atoms can be deexcited through collisions
As with Classic MW, atoms should not be deexcitable through collisions
immediately after being excited, but are instead stable for a certain
lifetime. After this lifetime, a collision can deexcite them either via
radiation or radiationless transmission.

Kept the stochastic method of determining spontaneous emission.

The stochastic method would not be easily useable for thermal transmission
because we would need to know the frequency of collisions.

[#58011196]