Hit should hide event from viewport drag handler
[#57815384]

By listening for viewport dragging on the <div> containing the svg/canvas sandwich layers, mouse events on atoms were bubbling to our viewport handler. This meant that dragging an atom with viewPortDragging = true dragged the atom across the model *and* dragged the model (so the atom went zooming at 2x speed).

By instead listening for viewport dragging events on the bottom-most <svg> of the sandwich, we get essentially the behavior we had in the svg-only view.