swallow, don't hit test, clicks on click shield
[#57656854]

Doing simulated svg hit testing using "sandwiched" svg/canvas layers forces us to simulate clicks -- browser-issued click events will be completely spurious because the browser doesn't know the real target of the corresponding mousedown and mouseup events, and so can't determine whether the mouseup and mousedown were targeted at the same element, or not.

Note also that even if the above did not apply, the only sensible thing for the canvas to do with click events dispatched to it is ignore them anyway. Only the canvas based view can tell if the same atom received mousedown and mouseup.