Unhide all layers when no element was hit during mouse hit test
This small issue wasn't visible, as in a typical scenario we always
hit background rectangle. However if you remove that rectangle (rect
svg element) and try to click empty area, you will notice that
everything will remain hidden. That small change fixes that.

It is useful for experiments with efficient 'mousemove' handling.