MD2D renderer: rm unused "atom tooltips" view code
This is not help tooltips, it's the old atom tooltips attached to each atom that were used to scaffold initial development of MD2D and are no longer actually used, at least as of 0.6.0 (PixiJS release).

Probably could implement simpler canvas-based version if we ever choose to reimplement atom tooltips.