Extract MD2D images handling code to ImagesHelper
[#57495138]

This should introduce no functional changes (functional improvements are what I want to do next). However, it's invaluable for clarifying precisely what parts of the MD2D renderer's internal state are required by the images-rendering code and what parts are not.

The pattern being used to share the MD2D renderer's closure variables with the ImagesHelper may seem a little unusual, but we need a way to share these variables with a friendly helper module by *name* rather than by value. These variables are initialized at different times and can be reassigned to, forcing you to remember to pass the new values to the helper module when they are updated, or to pass them as part of the argument list of every helper module method, or else to risk having stale or undefined values in the helper module.

If the MD2D renderer stored its state as properties of 'this' it would be much easier to share such variables by name with a helper module, and might make it easier to refactor large modules into smaller modules as is done here.