Implement VdW lines in Pixi
[#58090282]

In this implementation, all VdW lines have 5 "dash" segments (some of which are occluded by the atoms); they are stretched or shrunk horizontally to match the distance between the center of the atoms joined by the line. This is not only a faster and simpler implementation (allowing us to use sprites and just 1 texture), but the actual visual difference between this and the previous implementation, which used constant-width segments, is barely noticeable and not at all objectionable.