Draft "Excited States and Photons" interactives
[#56363566]

Bootstrapped from the the conversion-test "quantum-collision" and "quantum-emission" examples.

Also, this includes a custom model for Excited States and Photons page 3. The MML converter didn't include a light sources section in the automatically-converted model. (is this correct or not?) However, the quantum dynamics converter expects a light source property with an 'on' method.