Always include shutterbug even in static build
these changes make shutterbug available even in the static
distribution case. Currently shutterbug will not be
able to save an image because it won't have a valid server
to connect to. But if the static page is embedded in an iframe
inside of another page. Then the outer page can use its
shutterbug to take a picture of the frame.
[#59909576]