Chemical Reactions: fix energy conservation issues
[#60863274]

We were adding / removing new bonds, but weren't updating forces
after it. So atoms were moving for next 0.5dt not knowing that they
are already bonded.

If you take a look at TE e.g. in this interactive:
http://lab.dev.concord.org/interactives.html#interactives/conversion-tests/chemical-reactions/bonds-exchange.json
when heat bath is turned off, you can still notice some changes in TE.
However it seems to be caused by the imperfect bouncing off walls
algorithm that doesn't conserve energy when atoms are bonded.

Note that increases and drops in TE graph are visible when atoms
pair is close to the model boundary, not when reaction happens.