DataSet: Add loadDataSet and serialize functions
These move data in and out in exactly the same format as the
dataSet's internal _dataSeriesArry. It's possible that we may wish
to modify this so that the data is easier to parse by humans.

loadDataSet iterates through the data and calls appendStaticDataPoint
repeatedly for each row, allowing components that listen to SAMPLE_ADDED
to behave as if new data were coming in. We may want a quicker version
that just makes a direct copy of the array.

[#63473964]