Do not use listeningPool to listen to events on model object
[#63901650]

We can't call .removeAll() method of listeningPool here to, as the
same listeningPool is used to listen to events on dataSet (and we
do not want to remove dataSet listeners).

Perhaps we could create a new listening pool dedicated to model only.
But as this object is thrown away anyway, just used basic approach.