getVersionedUrl mv to common/url-helper.js. Add params for loading data.
[#64338690]

The methods for returning a versioned URL have been moved out to a helper.

The helper lets you add parameters to indicate things like try to load
learner data (so the initial interactive data isn’t loaded). And to allow
the interactive to display an warning / alert if the data won’t be saved.

https://www.pivotaltracker.com/story/show/64338690