Allow playbackSupport to be optional
IS [#63728770]

Mixing playbackSupport into a model causes the model's tick method to be called by an internal timer. However, this is not appropriate for models which "tick" on the receipt of external input, such as the sensor models. Because however these models can use every other part of the labModelerMixin, make just playbackSupport optional.