add sampleBatchLength property to signal generator
IS [#63728770]

This stores samples in a 'batch' of length sampleBatchSize, deferring all tick events until the batch is full. When the batch is full, it synchronously issues one tick per sample in the batch until the batch is emptied.

This will allow us to test the possible behavior of the LabQuest2 sensor model (which has to deal with receiving batches of samples from the LabQuest2) under the assumption that we maintain the "one tick per sample" assumption of the other models.