Batched signal generator interactive
IS [#63728770]

This is useful for testing how well the Lab grapher deals with receiving many synchronous ticks, each with a new sample. The first version of the LabQuest2 sensor model may receive batches of ~60 samples at a time, but the current architecture dictates that we issue on tick event for each sample (thus, 60 sequential tick events and 60 sequential draws). The dataset-based architecture which might avoid this issue is not quite ready yet.