Change CODAP exporter code to use formula for calculating row
Previously an auto generated column called “run” would be added to the run-level
outputs for data being exported to CODAP. This change attempts to use a formula
for the first column that CODAP will calculate (and recalculate if rows are deleted).
It also solves the issue of students reloading the page and having their “run” number
start back at 1.