Have all filters checked by default except for broken
With the new ability to search for interactives it has frequently required
that one or more filter checkboxes be checked after the page has been
loaded. Now all checkboxes are checked by default except for the broken
filter, so users may need to uncheck some if they want to find an interactive
by browsing the menu rather than by using the search box.