Use lab/mml-converter to perform MML conversion in node environment too
[#66517740]

The biggest difference is that now browsefied-cheerio is always used
(instead of node package).

It makes code reuse between node and browser versions simpler, but also
ensures that both methods return identical results. Otherwise it won't
be guaranteed. cheerio has to be browserified manually, including
some changes to its source, so in theory it can be a bit different
that node module.

Also md2d-node-api is updated.