use travis' built in bundler support
this does not use the --deployment option of bundler
this is because the Makefile expects at least `haml`
to work. That only works because any binaries that come
along with the gems are symlinked in a bin folder on the
PATH. When --deployment is used then the gems are 
installed in the vendor/bundle folder and the bins are
not symlinked.