there is no need for bundle install --binstubs
The one place it is still used is on the servers
They are misconfigured so that when capistrano runs tasks,
the shell isn't setup correctly so the gems are not
installed in the same place as when a user ssh'es into
the server. So the bin folder is still useful in that case.

[#68039776]