Prevent unnecessary re-requesting of svg <image>s
setImageProperties uses the 'update' path of the image renderer, which before this change would always touch the xlink:href attribute of <image> elements. Unfortunately, at least in Chrome, even setting the attribute to have the same value it already has can cause a request for that image. (Cache headers meliorate this problem on production and dev servers, but the Cache-Control header defaults to 'no-cache' on the local-development server.)