Fix bug when resizing playback controls to fit
[#79814322]

Turned out that PlaybackController._fitButtons was not idempotent; if a font-size was set in a previous call to the method, the font-size chosen by future calls could be affected (for the same dimensions and control layout)

This method is called several times during model setup.  In certain circumstances, early iterations set the playback controller font-size to its minimum of 10px, and this would affect future iterations such that they refused to reset the font-size to an appropriate, larger, value.

This fixes that particular problem by making sure to unset the font-size before making size calculations that affect the final font-size.

On the other hand, control resizing is still something of a hack.