Add setting to skip PE check when adding or dragging atoms
This is a new top-level property, skipPECheckOnAddAtom, default=false.

This is used every time model.addAtom is called (e.g. by a script), and
when an atom is dragged by a user when the model is stopped.

It is slightly finer-grained than the existing "checkLocation" option
that is used in setAtomProperties, as it only skips PE calculation,
not obstacle- or border-checking.

This is for the bond-creation models [#74687984] and [#74688264]

[#82674410]