Make text not clickable if it's attached to an atom or obstacle
Text boxes currently capture clicks to allow an author dragging while in
"authoring mode." It is not clear whether authoring mode is still used,
but in any case, this dragging would be pointless for atoms attached to
objects. Therefore, this does not affect the current behavior.

This allows us to have large element symbols in [#74688264]

[#82686348]